requirements.txt
Flask==2.2.2
pytube==15.0.0
pytubefix==8.12.2
streamlit==1.43.1
elevenlabs
python-dotenv
altair==4.2.0
brave-search==0.1.8
rank_bm25 == 0.2.2
openai==1.42.0
ruamel.yaml==0.18.10
Werkzeug==2.3.8
PyYAML==6.0.2
faiss-cpu==1.8.0.post1